Split AtomStorage into ODBC and PostgreSQL
