+++ opencog/atomspace/AtomSpace.h
@@ -56,7 +56,8 @@ namespace opencog
 class AtomSpace
 {
     friend class Atom;               // Needs to call get_atomtable()
-    friend class SQLPersistSCM;
+    friend class AtomStorage;
+    friend class BackingStore;
     friend class ZMQPersistSCM;
     friend class ::AtomTableUTest;
