+++ tests/persist/sql/BasicSaveUTest.cxxtest
@@ -35,7 +35,7 @@
 #include <opencog/truthvalue/SimpleTruthValue.h>
 #include <opencog/atomspace/TLB.h>
 #include <opencog/atoms/base/atom_types.h>
-#include <opencog/persist/sql/AtomStorage.h>
+#include <opencog/persist/sql/ODBCAtomStorage.h>
 
 #include <opencog/util/Logger.h>
 #include <opencog/util/Config.h>
@@ -121,7 +121,7 @@ class BasicSaveUTest :  public CxxTest::TestSuite
 			exit(1);
 		}
 
-		AtomStorage *astore;
+		ODBCAtomStorage *astore;
 		void single_atom_save_restore(std::string id);
 		void add_to_table(int, AtomTable *, std::string id);
 		void check_table(int, AtomTable *, std::string id);
@@ -137,7 +137,7 @@ void BasicSaveUTest::setUp(void)
 {
 	as = new AtomSpace();
 
-	astore = new AtomStorage(dbname, username, passwd);
+	astore = new ODBCAtomStorage(dbname, username, passwd);
 	if (!astore->connected())
 	{
 		logger().info("setUp: cannot connect to database");
@@ -214,7 +214,7 @@ static void atomCompare(AtomPtr a, AtomPtr b, const char* where)
  */
 void BasicSaveUTest::single_atom_save_restore(std::string id)
 {
-	AtomStorage *store = new AtomStorage(dbname, username, passwd);
+	ODBCAtomStorage *store = new ODBCAtomStorage(dbname, username, passwd);
 	if (!store->connected())
 	{
 		logger().debug("single_atom_save_restore: cannot connect to db");
@@ -350,7 +350,7 @@ void BasicSaveUTest::test_table(void)
 {
 	logger().debug("BEGIN TEST: %s", __FUNCTION__);
 
-	AtomStorage *store = new AtomStorage(dbname, username, passwd);
+	ODBCAtomStorage *store = new ODBCAtomStorage(dbname, username, passwd);
 	if (!store->connected())
 	{
 		logger().debug("test_table: cannot connect to db");
@@ -371,7 +371,7 @@ void BasicSaveUTest::test_table(void)
 	delete table1;
 
 	// Reopen connection, and load the atom table.
-	store = new AtomStorage(dbname, username, passwd);
+	store = new ODBCAtomStorage(dbname, username, passwd);
 	TSM_ASSERT("Not connected to database", store->connected());
 
 	AtomTable *table2 = new AtomTable();
